pyproject.toml
[project]
name = "preprocessor"
version = "0.0.1"
description = "Benthic image preprocessor"
keywords = ["preprocessor"]
license-expression = "Apache-2.0"
license-file = "LICENSE.md"
license-files = ["LICENSE.md"]
readme = {file = "README.md", content-type = "text/markdown"}
authors = [
  {name = "Daniel A. A. Pelsmaeker", email = "developer@pelsmaeker.net"},
]
maintainers = [
  {name = "Daniel A. A. Pelsmaeker", email = "developer@pelsmaeker.net"},
]
requires-python = ">=3.12"
dependencies = [
  "click>=8.3.1",
  "numpy>=2.2.6",
  "opencv-contrib-python-headless==4.12.0.88",  # For OpenCV version that works on MacOS
  "pillow>=12.0.0",
  "pydantic>=2.12.5",
  "PySide6-Essentials>=6.10.1",
#  "pyside6>=6.10.1",
  "qimage2ndarray>=1.10.0",
  "scikit-learn>=1.7.2",
]

[project.urls]
Homepage = "https://cambiomed-biodiversa.com/"
#Documentation = "https://readthedocs.org"
Repository = "https://github.com/CAMBioMed/preprocessor.git"
Source = "https://github.com/CAMBioMed/preprocessor/"
Issues = "https://github.com/CAMBioMed/preprocessor/issues"
Changelog = "https://github.com/CAMBioMed/preprocessor/blob/main/CHANGELOG.md"


# CLI launcher
[project.scripts]
preprocessor = "preprocessor.main:main"

# For Windows GUI launcher (no console)
[project.gui-scripts]
preprocessor-gui = "preprocessor.main:main_gui"

[dependency-groups]
# NOTE: This list must be kept in sync with `tool.briefcase.app.preprocessor.test_requires` test dependencies.
# Please update both lists when adding/removing dependencies, until UV is supported by Briefcase.
dev = [
  # Test dependencies
  "pytest>=9.0.1",
  "pytest-cov>=7.0.0",
  "pytest-qt>=4.5.0",
  # Build dependencies
  "hatch>=1.16.1",
  "hatchling>=1.28.0",
  # Development dependencies
  # For PyCharm 'Show as Image'
  "mypy>=1.19.0",
  "pillow>=12.0.0",
  "pyright>=1.1.406",
  "ruff>=0.14.7",
  "yamlfix>=1.19.0",
]


################
# Build System #
################

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
  "src/preprocessor",
]


##########
# PyTest #
##########

[tool.pytest.ini_options]
minversion = "6.0"
addopts = ["-ra", "--showlocals", "--strict-markers", "--strict-config"]
xfail_strict = true
filterwarnings = [
  "error",
]
log_cli_level = "INFO"
testpaths = [
  "tests",
]


############
# Coverage #
############

[tool.coverage]
run.source = ["preprocessor"]
port.exclude_lines = [
  'pragma: no cover',
  '\.\.\.',
  'if typing.TYPE_CHECKING:',
]


########
# MyPy #
########

[tool.mypy]
warn_return_any = true
disallow_untyped_defs = true


#############################
# Ruff Linting & Formatting #
#############################

[tool.ruff]
line-length = 120
exclude = [
    "ui_*.py",
    "tests",
]

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = [
  # flake8-2020
  "YTT",
  # flake8-annotations
  "ANN",
  # flake8-bugbear
  "B",
  # flake8-comprehensions
  "C4",
  # flake8-errmsg
  "EM",
  # flake8-executable
  "EXE",
  # flake8-implicit-str-concat
  "ISC",
  # flake8-import-conventions
  "ICN",
  # flake8-logging-format
  "G",
  # flake8-pie
  "PIE",
  # flake8-pytest-style
  "PT",
  # flake8-return
  "RET",
  # flake8-simplify
  "SIM",
  # flake8-unused-arguments
  "ARG",
  "E",
  # pycodestyle
  "W",
  # Pyflakes
  "F",
  # pydocstyle
  "D",
  # pylint
  "PL",
  # pyupgrade
  "UP",
  # Ruff-specific
  "RUF",
]
ignore = [
  # Stylistic choices
  # Pylint/Refactor
  "PLR",
  # unnecessary-assign
  "RET504",
  # Ensure formatter compatibility
  # whitespace-before-punctuation
  "E203",
  # Allow gradual documentation
  # undocumented-public-module
  "D100",
  # undocumented-public-class
  "D101",
  # undocumented-public-method
  "D102",
  # undocumented-public-nested-class
  "D103",
  # undocumented-public-package
  "D104",
  # undocumented-magic-method
  "D105",
  # undocumented-public-nested-class
  "D106",
  # undocumented-public-init
  "D107",
  # missing-blank-line-after-summary
  "D205",
  # non-imperative-mood
  "D401",
  # import-outside-top-level
  "PLC0415",
  # module-import-not-at-top-of-file
  "E402",
  # For now:
  # logging-f-string
  "G004",
]
pydocstyle.convention = "pep257"

[tool.ruff.lint.per-file-ignores]
# Tests can ignore a few extra rules
"tests/**.py" = [
    # Missing return type annotation for public function
    "ANN201",
    # Missing docstring
    "D103",
    # Use of assert is detected
    "S101",
    # Missing __init__.py file
    "INP001",
]

[tool.ruff.lint.flake8-annotations]
allow-star-arg-any = true


###########
# YamlFix #
###########
[tool.yamlfix]
line_length = 120
sequence_style = "keep_style"
explicit_start = false
whitelines = 1
section_whitelines = 1


##############
# Qt Creator #
##############
[tool.pyside6-project]
files = [
    "src/preprocessor/__main__.py",
    "src/preprocessor/gui/editor_dock.ui",
    "src/preprocessor/gui/export_dialog.ui",
    "src/preprocessor/gui/hsv_threshold.ui",
    "src/preprocessor/gui/launch_dialog.ui",
    "src/preprocessor/gui/main.ui",
    "src/preprocessor/gui/properties_dock.ui",
]


#############
# Briefcase #
#############
[tool.briefcase]
project_name = "CAMBioMed Preprocessor"
bundle = "com.cambiomed-biodiversa"
license.file = "LICENSE.md"
author = "Daniel A. A. Pelsmaeker"
author_email = "developer@pelsmaeker.net"

[tool.briefcase.app.preprocessor]
formal_name = "CAMBioMed Preprocessor"
long_description = """Preprocessor application for benthic photos."""
icon = "appicon/preprocessor"
splash_background_color = "#a95e43"
sources = [
    "src/preprocessor",
]
test_sources = [
    "tests",
]

requires = [
    # Add your cross-platform app requirements here
]
# NOTE: This list must be kept in sync with `dependency-groups.dev` test dependencies.
# Please update both lists when adding/removing dependencies, until UV is supported by Briefcase.
# See also: https://github.com/beeware/briefcase/issues/1235
test_requires = [
  # Test dependencies
  "pytest>=9.0.1",
  "pytest-cov>=7.0.0",
  "pytest-qt>=4.5.0",
]

[tool.briefcase.app.preprocessor.macOS]
# WORKAROUND: https://github.com/beeware/briefcase/issues/2658#issuecomment-3832236695
min_os_version = "13.0"
universal_build = true      # Support both x86_64/ARM64
requires = [
    "std-nslog~=1.0.3",
]

[tool.briefcase.app.preprocessor.linux]
requires = []

[tool.briefcase.app.preprocessor.linux.system.debian]
system_requires = []

system_runtime_requires = [
    # Derived from https://doc.qt.io/qt-6/linux-requirements.html
    "libxext6",
    "libxrender1",
    "libx11-xcb1",
    "libxkbcommon-x11-0",
    "libxcb-image0",
    "libxcb-cursor0",
    "libxcb-shape0",
    "libxcb-randr0",
    "libxcb-xfixes0",
    "libxcb-sync1",
    "libxcb-icccm4",
    "libxcb-keysyms1",
    "libfontconfig1",
    "libsm6",
    "libice6",
    "libglib2.0-0",
    "libgl1",
    "libegl1",
    "libdbus-1-3",
]

[tool.briefcase.app.preprocessor.linux.system.rhel]
system_requires = []

system_runtime_requires = [
    "qt6-qtbase-gui",
]

[tool.briefcase.app.preprocessor.linux.system.suse]
system_requires = []

system_runtime_requires = [
    "libgthread-2_0-0",
    "libQt6Gui6",
]

[tool.briefcase.app.preprocessor.linux.system.arch]
system_requires = []

system_runtime_requires = [
    "qt6-base",
]

[tool.briefcase.app.preprocessor.linux.appimage]
manylinux = "manylinux_2_28"

system_requires = [
# ?? FIXME
]

linuxdeploy_plugins = [
]

[tool.briefcase.app.preprocessor.linux.flatpak]
flatpak_runtime = "org.kde.Platform"
flatpak_runtime_version = "6.9"
flatpak_sdk = "org.kde.Sdk"

[tool.briefcase.app.preprocessor.windows]
requires = []

# Mobile deployments
[tool.briefcase.app.preprocessor.iOS]
supported = false

[tool.briefcase.app.preprocessor.android]
supported = false

# Web deployments
[tool.briefcase.app.preprocessor.web]
supported = false
